@@ -28,6 +28,7 @@
     "sqlite3": "^5.0.2"
   },
   "devDependencies": {
-    "eslint": "^6.8.0"
-  }
+    "eslint": "^8.9.0"
+  },
+  "type": "module"
 }
