@@ -1,5 +1,8 @@
 import sqlite3 from "sqlite3";
-import path from "path";
+import path, { dirname } from "path";
+import { fileURLToPath } from "url";
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 class WeathermanDAO {
   constructor() {
