@@ -1,4 +1,4 @@
-import WeathermanDAO from "../dao";
+import WeathermanDAO from "../dao.js";
 import cron from "node-cron";
 
 export const clearMonthlyLogs = () => {
