@@ -1,7 +1,7 @@
 import cron from "node-cron";
-import WeathermanDAO from "../../server/dao";
-import { ShuffleRandomizer } from "../utils/random-utils";
-import { configStore, KEYS } from "../../config";
+import WeathermanDAO from "../../server/dao.js";
+import { ShuffleRandomizer } from "../utils/random-utils.js";
+import { configStore, KEYS } from "../../config/index.js";
 
 const OKR_LIST = [
   "Protect Brand Reputation & Excel in Operational Excellence",
